Added options to specify character spacings, and also which character sets to use. Using anything other than ascii tends to break things though,
